Convert to UTF-8 before printing (fixes #31)
Python tries to identify if the stdout is being written to a device
capable of displaying UTF-8 characters or not.
The terminal is detected as UTF-8 capable whereas it tries to
convert to ASCII when writing to files. This is described here:
http://blog.mathieu-leplatre.info/python-utf-8-print-fails-when-redirecting-stdout.html

I'll always forceful convert to UTF-8 anyways to avoid any error.